getGithubUser to TS
