+++ backend/src/serverHelper.js
@@ -8,7 +8,6 @@ const { XMLHttpRequest } = require('xmlhttprequest');
 const path = require('path');
 const fetch = require('node-fetch');
 const AdmZip = require('adm-zip');
-const passport = require('passport');
 const os = require('os');
 const mongo = require('./database/DbServices');
 const emptyScenario = require('./models/emptyScenario');
@@ -449,49 +448,7 @@ async function updateScenarioTestStatus(uploadedReport) {
 	}
 }
 
-const getGithubData = (res, req, accessToken) => {
-	fetch(
-		`https://api.github.com/user?access_token=${accessToken}`,
-		{
-			uri: `https://api.github.com/user?access_token=${accessToken}`,
-			method: 'GET',
-			headers:
-			{
-				'User-Agent': 'SampleOAuth',
-				Authorization: `Token ${accessToken}`
-			}
-		}
-	)
-		.then((response) => response.json())
-		.then(async (json) => {
-			console.log('JSON in GetGitHubData');
-			console.log(json);
-			req.body = json;
-			req.body.githubToken = accessToken;
-			try {
-				await mongo.findOrRegisterGithub(req.body);
-				passport.authenticate('github-local', (error, user) => {
-					if (error || !user) res.json({ error: 'Authentication Error' });
-					req.logIn(user, (LoginError) => {
-						if (LoginError) {
-							res.json({ error: 'Login Error' });
-						} else {
-							res.header('Access-Control-Allow-Origin', process.env.FRONTEND_URL || 'http://localhost:4200');
-							res.header('Access-Control-Allow-Credentials', 'true');
-							res.header('Access-Control-Allow-Headers', 'Origin, X-Requested-With, Content-Type, Accept, Credentials');
-							res.json({
-								login: user.github.login,
-								id: user.github.id
-							});
-						}
-					});
-				})(req, res);
-			} catch (error) {
-				console.log('getGithubData error:', error);
-				res.sendStatus(400);
-			}
-		});
-};
+
 
 async function exportSingleFeatureFile(_id, source) {
 	const dbStory = mongo.getOneStory(_id, source);
@@ -520,7 +477,6 @@ module.exports = {
 	updateLatestTestStatus,
 	getReportHistory,
 	uniqueRepositories,
-	getGithubData,
 	cleanFileName,
 	getFeatureContent,
 	getScenarioContent,
